+++ examples/Registration/smsRegistration.php
@@ -2,6 +2,14 @@
 
 require __DIR__.'/../../vendor/autoload.php';
 
+class ExtendedInstagram extends \InstagramAPI\Instagram
+{
+    public function sendLoginFlow()
+    {
+        $this->_sendLoginFlow(true);
+    }
+}
+
 /////////////////
 $debug = true;
 /////////////////
@@ -16,8 +24,11 @@ $year = 1970;
 $firstName = 'First Name';
 /////////////////
 
-$ig = new \InstagramAPI\Instagram($debug);
+$ig = new ExtendedInstagram($debug);
 $ig->setUserWithoutPassword($username);
+
+$ig->event->updateAppState('foreground', 'not_initialized');
+$ig->event->sendFlowSteps('landing', 'sim_card_state', $waterfallId, $startTime, ['flow' => 'phone']);
 
 $launcherResponse = $ig->internal->sendLauncherSync(true)->getHttpResponse();
 $ig->settings->set('public_key', $launcherResponse->getHeaderLine('ig-set-password-encryption-pub-key'));
@@ -33,6 +44,8 @@ $ig->event->sendFlowSteps('one_page_v2', 'step_view_loaded', $waterfallId, $star
 
 $ig->account->checkPhoneNumber($phone);
 
+$ig->event->sendFlowSteps('phone', 'next_button_tapped', $waterfallId, $startTime, ['flow' => 'phone']);
+
 $tos = $ig->account->requestRegistrationSms($phone, $waterfallId, $username)->getTosVersion();
 echo 'SMS Code: ';
 $smsCode = trim(fgets(STDIN));
@@ -45,24 +58,40 @@ $currentYear = date('Y');
 $currentMonth = date('m');
 $currentDay = date('d');
 
+usleep(mt_rand(1000000, 3000000));
 for ($i = 1; $i <= ($currentYear - $year); $i++) {
     $ig->event->sendDobPick(sprintf('%d-%02d-%02d', $currentYear - $i, $currentMonth, $currentDay));
+    usleep(mt_rand(1000, 3000));
 }
 
+usleep(mt_rand(1000000, 3000000));
 $operand = ($month > $currentMonth) ? 1 : -1;
 for ($i = 1; $i <= abs(($currentMonth - $month)); $i++) {
     $ig->event->sendDobPick(sprintf('%d-%02d-%02d', $year, $currentMonth + ($operand * $i), $currentDay));
+    usleep(mt_rand(1000, 3000));
 }
 
+usleep(mt_rand(1000000, 3000000));
 $operand = ($day > $currentDay) ? 1 : -1;
 for ($i = 1; $i <= abs(($currentDay - $day)); $i++) {
     $ig->event->sendDobPick(sprintf('%d-%02d-%02d', $year, $month, $currentDay + ($operand * $i)));
+    usleep(mt_rand(1000, 3000));
 }
 
 $response = $ig->internal->checkAgeEligibility($day, $month, $year);
 if ($response->getEligibleToRegister() !== true) {
     exit();
 }
+
+$ig->event->sendFlowSteps('username', 'step_view_loaded', $waterfallId, $startTime,
+    [
+        'is_facebook_app_installed' => false,
+        'messenger_installed'       => false,
+        'whatsapp_installed'        => false,
+        'fb_lite_installed'         => false,
+        'flow'                      => 'phone',
+    ]
+);
 
 $ig->event->sendFlowSteps('username', 'step_view_loaded', $waterfallId, $startTime, ['flow' => 'phone']);
 $ig->event->sendNavigation('button', 'add_birthday', 'username_sign_up');
@@ -73,9 +102,10 @@ $ig->event->sendFlowSteps('username', 'next_button_tapped', $waterfallId, $start
 $ig->event->sendFlowSteps('username', 'register_with_ci_option', $waterfallId, $startTime, ['flow' => 'phone']);
 
 $ig->internal->startNewUserFlow();
-$ig->internal->getOnBoardingSteps($waterfallId);
+$ig->internal->getOnBoardingSteps($waterfallId, 'phone');
 
-sleep(mt_rand(5, 10));
+$ig->event->forceSendBatch();
+usleep(mt_rand(15000000, 3000000));
 
 try {
     $response = $ig->account->createValidated($smsCode, $username, $password, $phone, sprintf('%d-%2d-%2d', $year, $month, $day), $firstName, $waterfallId, $tos);
@@ -83,6 +113,98 @@ try {
     echo 'Something went wrong: '.$e->getMessage()."\n";
     exit();
 }
+
+// Sets active status
+$ig->isMaybeLoggedIn = true;
+$ig->account_id = $response->getCreatedUser()->getPk();
+$ig->settings->set('account_id', $ig->account_id);
+$ig->settings->set('last_login', time());
+
 $ig->event->sendFlowSteps('done', 'register_account_request_submitted', $waterfallId, $startTime, ['flow' => 'phone']);
 $ig->event->sendFlowSteps('done', 'register_account_created', $waterfallId, $startTime, ['flow' => 'phone', 'instagram_id' => $response->getUser()->getPk()]);
 $ig->event->forceSendBatch();
+
+$ig->account->getAccountFamily();
+$ig->internal->getMobileConfig(false);
+$ig->event->sendZeroCarrierSignal();
+
+$ig->internal->getOnBoardingSteps($waterfallId, 'phone');
+
+$feed = $this->timeline->getTimelineFeed(null, [
+    'reason' => Constants::REASONS[0],
+]);
+
+$followingList = json_decode($ig->discover->getAyml('explore_people')->getMaxId(), true);
+$ig->people->getFriendships($followingList);
+
+$userToFollow = array_rand($followingList);
+$ig->event->sendFollowButtonTapped($userToFollow, 'discover_people_nux',
+    [
+        [
+            'module'        => 'username_sign_up',
+            'click_point'   => 'button',
+        ],
+        [
+            'module'        => 'username_sign_up',
+            'click_point'   => 'button',
+        ],
+        [
+            'module'        => 'register_flow_add_profile_photo',
+            'click_point'   => 'button',
+        ],
+        [
+            'module'        => 'username_sign_up',
+            'click_point'   => 'button',
+        ],
+        [
+            'module'        => 'add_birthday',
+            'click_point'   => 'button',
+        ],
+        [
+            'module'        => 'one_page_registration',
+            'click_point'   => 'button',
+        ],
+        [
+            'module'        => 'phone_confirmation',
+            'click_point'   => 'button',
+        ],
+    ]
+);
+$ig->people->follow($userToFollow);
+$ig->event->sendProfileAction('follow', $userToFollow,
+    [
+        [
+            'module'        => 'username_sign_up',
+            'click_point'   => 'button',
+        ],
+        [
+            'module'        => 'username_sign_up',
+            'click_point'   => 'button',
+        ],
+        [
+            'module'        => 'register_flow_add_profile_photo',
+            'click_point'   => 'button',
+        ],
+        [
+            'module'        => 'username_sign_up',
+            'click_point'   => 'button',
+        ],
+        [
+            'module'        => 'add_birthday',
+            'click_point'   => 'button',
+        ],
+        [
+            'module'        => 'one_page_registration',
+            'click_point'   => 'button',
+        ],
+        [
+            'module'        => 'phone_confirmation',
+            'click_point'   => 'button',
+        ],
+    ]
+);
+
+$rankToken = \InstagramAPI\Signatures::generateUUID();
+$ig->event->sendSearchFollowButtonClicked($userToFollow, 'discover_people_nux', $rankToken);
+
+$ig->sendLoginFlow();